+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Dropped CSS reset and Lato font in favor of a full Chakra theme (#159)
 -   Include key in filename when downloading a model (#160)
 -   Made /compute_plans the landing page (#161)
+-   New visual style for navigation on compute plan page (#158)
 
 ## [0.3.0] - 2021-10-04